refactor(workspace): restructured workspace
Subdirectories now match actual package names.
