+++ .gitignore
@@ -155,6 +155,7 @@ Temporary Items
 
 # End of https://www.gitignore.io/api/osx,node,intellij
 
-*/lib
-*/reports/
-!*/reports/.gitkeep
+*/*/dist
+*/*/lib
+*/*/reports/
+!*/*/reports/.gitkeep
