+++ @kangaroo/angular-oauth2/package.json
