+++ @kangaroo/angular-oauth2/src/admin/index.ts
