+++ @kangaroo/angular-oauth2/src/index.spec.ts
