+++ @kangaroo/angular-oauth2/src/index.ts
