+++ @kangaroo/angular-oauth2/src/oauth2/index.ts
