+++ @kangaroo/angular-oauth2/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tsconfig.json",
+  "extends": "../../tsconfig.json",
 
   "compilerOptions": {
     "sourceRoot": "./src",
