+++ @kangaroo/authz-ui/src/app/app.component.spec.ts
