+++ @kangaroo/authz-ui/src/app/app.component.ts
