+++ @kangaroo/authz-ui/src/app/app.module.ts
