+++ @kangaroo/authz-ui/src/environments/environment.prod.ts
