+++ @kangaroo/authz-ui/src/favicon.ico
