+++ @kangaroo/authz-ui/src/polyfills.ts
