+++ @kangaroo/authz-ui/src/styles.css
