+++ @kangaroo/authz-ui/src/typings.d.ts
