+++ @kangaroo/jwt/karma.conf.js
@@ -15,20 +15,15 @@
  * limitations under the License.
  */
 
-// Karma configuration file, see link for more information
-// https://karma-runner.github.io/1.0/config/configuration-file.html
-
-const commonConfig = require('../common/karma.conf');
+const commonConfig = require('../../common/karma.conf');
 
 module.exports = function (config) {
-
     config.set(commonConfig);
-
     config.set({
-        basePath: '',
-        frameworks: ['jasmine'],
+        frameworks: ['jasmine', 'karma-typescript'],
         plugins: [
             require('karma-jasmine'),
+            require('karma-typescript'),
 
             require('karma-chrome-launcher'),
             require('karma-webdriver-launcher'),
@@ -36,8 +31,12 @@ module.exports = function (config) {
 
             require('karma-junit-reporter'),
             require('karma-coverage-istanbul-reporter'),
-            require('karma-spec-reporter')
+            require('karma-spec-reporter'),
         ],
-        files: ['src/**/*.spec.ts']
+        files: ['src/**/*.ts'],
+        preprocessors: {
+            'src/**/*.ts': ['karma-typescript'],
+        },
+        reporters: ['spec', 'junit', 'karma-typescript']
     });
 };
