+++ @kangaroo/jwt/src/index.spec.ts
