+++ @kangaroo/jwt/src/json-web-token.ts
