+++ @kangaroo/jwt/src/jwt.spec.ts
