+++ @kangaroo/jwt/src/text-util.spec.ts
