+++ @kangaroo/jwt/src/text-util.ts
