--- kangaroo-angular-oauth2/tslint.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../tslint.json"
-}
